pyproject.toml

[project]
name = ""
version = ""
description = ""
authors = [
    {name = "J Stuff", email = "68690676+J-Stuff@users.noreply.github.com"},
]
dependencies = [
    "aiohttp==3.8.5",
    "aiosignal==1.3.1",
    "asn1crypto==1.5.1",
    "async-timeout==4.0.2",
    "asyncio==3.4.3",
    "asyncio-dgram==2.1.2",
    "attrs==23.1.0",
    "betterproto==2.0.0b6",
    "certifi==2023.7.22",
    "cffi==1.15.1",
    "charset-normalizer==3.2.0",
    "cryptography==41.0.3",
    "deprecation==2.1.0",
    "discord-py==2.3.1",
    "dnspython==2.4.1",
    "frozenlist==1.4.0",
    "grpclib==0.4.5",
    "h2==4.1.0",
    "hpack==4.0.0",
    "http-ece==1.1.0",
    "hyperframe==6.0.1",
    "idna==3.4",
    "mcstatus==11.0.0",
    "multidict==6.0.4",
    "numpy==1.25.2",
    "oscrypto==1.3.0",
    "packaging==23.1",
    "Pillow==10.0.0",
    "protobuf==4.23.4",
    "pycparser==2.21",
    "python-dateutil==2.8.2",
    "python-dotenv==1.0.0",
    "rcon==2.4.2",
    "requests==2.31.0",
    "scipy==1.9.3",
    "six==1.16.0",
    "steamquery==1.0.2",
    "timeout-decorator==0.5.0",
    "urllib3==2.0.4",
    "websockets==11.0.3",
    "yarl==1.9.2",
    "termcolor>=2.4.0",
    "schedule>=1.2.1",
    "steam>=1.4.4",
    "flask>=3.0.2",
    "Flask-Discord>=0.1.69",
    "schedule==1.2.0",
    "termcolor==2.4.0",
]
requires-python = ">=3.11"
license = {text = "MIT"}
readme = "README.md"


[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

